+++ src/sox_private.h
@@ -8,8 +8,8 @@ extern uint16_t *g_celVol;                  // mv           *1
 extern int16_t *g_celTmp;                   // dregree      *10
 extern uint16_t *g_celSOC;                  // %            *10
 extern uint16_t *g_grpSOC;                  // %            *10
-extern double  *g_celSOH;                   // %            *1
-extern double  *g_grpSOH;                   // %            *1
+extern float  *g_celSOH;                   // %            *1
+extern float  *g_grpSOH;                   // %            *1
 extern uint32_t *g_cycleCount;              // times        *1000
 extern float *g_accChgWH;                   // WH           *1
 extern float *g_accDsgWH;                   // WH           *1
